requirements.txt
aniso8601==9.0.1
biopython==1.79               
blinker==1.4                   
click==8.0.1                 
cssselect==1.1.0              
cycler==0.10.0                            
et-xmlfile==1.1.0                
ete3==3.1.2                 
flask==2.0.1                 
flask-mail==0.9.1               
flask-restful==0.3.9               
flask-sockets==0.2.1            
gevent==21.8.0             
gevent-websocket==0.10.1       
greenlet==1.1.1               
helperlibs==0.2.1        
itsdangerous==2.0.1            
jdcal==1.4.1                 
jinja2==3.0.1                 
lxml==4.6.3                
markupsafe==2.0.1
matplotlib==3.3.4
networkx==2.6.2           
numpy==1.20.2       
openpyxl==3.0.7              
pillow==8.3.1                 
pyexcelerator==0.6.4.1          
pyparsing==2.4.7              
pypng==0.0.20             
pyquery==1.4.3              
pysvg-py3==0.2.2.post3               
python-dateutil==2.8.2             
pytz==2021.1
pyuwsgi==2.0.21               
redis==3.5.3                 
redis-server==5.0.7 
six==1.16.0                
straight-plugin==1.5.0                 
werkzeug==2.0.1
                    
